Add files via upload
This project involves an in-depth analysis of the Diabetes dataset obtained from Kaggle. The dataset provides information on various factors related to diabetes, such as glucose levels, blood pressure, body mass index (BMI), and more. The objective of this analysis is to gain insights into the dataset and explore the relationships between different variables to better understand the risk factors associated with diabetes.
The project is implemented in Python, utilizing popular data analysis libraries such as Pandas, NumPy, and Scikit-learn. The code, along with the dataset, is provided in this GitHub repository to facilitate transparency, reproducibility, and collaboration.
Feel free to explore the repository and provide feedback or suggestions for further improvement. Your contributions and insights are highly appreciated!